Attempt at cleaning up the socket work
If only one call is made when closing mGBA-http, mGBA via the script throws an "unknown error". If mGBA-http is closed after more than one call, it's mostly "disconnected" that comes from the script. To differentiate the two for debugging purposes, an asterisks is added for the "unknown error" log line.