Wrote a basic load tester to find more socket problems
In the end, with enough requests (such as from the integration tests) the socket work would encounter socket reuse problems, or invalid argument issues. To address this, some extra resilience was written in the socket work, and the SocketService lifetime moved from singleton to scoped. Which now passes the load balancer test and the integration tests. Though the code is much messier and needs a cleanup.

I assume users wouldn't run 10s of requests per second, but that shouldn't be out of the expectations for the code to handle.